# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Authula API",
    "description": "Authula API - An open-source authentication solution that scales with you.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/organizations": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "summary": "List organizations",
        "description": "Lists every organization the authenticated user can access, both the ones they own and the ones they are a member of, newest first. Results are paginated: `page` defaults to 1 and `limit` defaults to 10, with a hard maximum of 100. Out-of-range values are clamped silently rather than rejected.",
        "operationId": "listOrganizations",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListOrganizationsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Organizations"
        ],
        "summary": "Create organization",
        "description": "Creates a new organization with the authenticated user as the owner.",
        "operationId": "createOrganization",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          }
        }
      }
    },
    "/organizations/{organization_id}": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "summary": "Get organization",
        "description": "Retrieves an organization by its ID.",
        "operationId": "getOrganization",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Organizations"
        ],
        "summary": "Delete organization",
        "description": "Deletes an organization.",
        "operationId": "deleteOrganization",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteOrganizationResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Organizations"
        ],
        "summary": "Update organization",
        "description": "Updates an existing organization.",
        "operationId": "updateOrganization",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganizationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          }
        }
      }
    },
    "/organizations/{organization_id}/invitations": {
      "get": {
        "tags": [
          "Organization Invitations"
        ],
        "summary": "List invitations",
        "description": "Lists the invitations for an organization, newest first. Results are paginated: `page` defaults to 1 and `limit` defaults to 10, with a hard maximum of 100. Out-of-range values are clamped silently rather than rejected.",
        "operationId": "listOrganizationInvitations",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListOrganizationInvitationsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Organization Invitations"
        ],
        "summary": "Create invitation",
        "description": "Creates an invitation for a user to join an organization.",
        "operationId": "createOrganizationInvitation",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "redirect_url",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationInvitationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationInvitation"
                }
              }
            }
          }
        }
      }
    },
    "/organizations/{organization_id}/invitations/{invitation_id}": {
      "get": {
        "tags": [
          "Organization Invitations"
        ],
        "summary": "Get invitation",
        "description": "Retrieves a single invitation.",
        "operationId": "getOrganizationInvitation",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "invitation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetOrganizationInvitationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/organizations/{organization_id}/invitations/{invitation_id}/accept": {
      "patch": {
        "tags": [
          "Organization Invitations"
        ],
        "summary": "Accept invitation",
        "description": "Accepts an invitation to join an organization.",
        "operationId": "acceptOrganizationInvitation",
        "parameters": [
          {
            "name": "redirect_url",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "invitation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptOrganizationInvitationQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationInvitation"
                }
              }
            }
          }
        }
      }
    },
    "/organizations/{organization_id}/invitations/{invitation_id}/reject": {
      "patch": {
        "tags": [
          "Organization Invitations"
        ],
        "summary": "Reject invitation",
        "description": "Rejects an invitation to join an organization.",
        "operationId": "rejectOrganizationInvitation",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "invitation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationInvitation"
                }
              }
            }
          }
        }
      }
    },
    "/organizations/{organization_id}/invitations/{invitation_id}/revoke": {
      "patch": {
        "tags": [
          "Organization Invitations"
        ],
        "summary": "Revoke invitation",
        "description": "Revokes a pending invitation.",
        "operationId": "revokeOrganizationInvitation",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "invitation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationInvitation"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcceptOrganizationInvitationQuery": {
        "properties": {
          "redirect_url": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object"
      },
...
      "CreateOrganizationInvitationRequest": {
        "properties": {
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "role"
        ],
        "type": "object"
      },
      "CreateOrganizationRequest": {
        "properties": {
          "logo": {
            "type": [
              "null",
              "string"
            ]
          },
          "metadata": {
            "additionalProperties": {},
            "type": [
              "object",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "slug": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "name",
          "role"
        ],
        "type": "object"
      },
...
      "DeleteOrganizationResponse": {
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "type": "object"
      },
...
      "GetOrganizationInvitationResponse": {
        "properties": {
          "invitation": {
            "$ref": "#/components/schemas/OrganizationInvitation"
          },
          "organization": {
            "$ref": "#/components/schemas/OrganizationSummary"
          }
        },
        "required": [
          "invitation",
          "organization"
        ],
        "type": "object"
      },
...
      "ListOrganizationInvitationsResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/GetOrganizationInvitationResponse"
            },
            "type": "array"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "required": [
          "data",
          "pagination"
        ],
        "type": "object"
      },
...
      "ListOrganizationsResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Organization"
            },
            "type": "array"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "required": [
          "data",
          "pagination"
        ],
        "type": "object"
      },
...
      "Organization": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "logo": {
            "type": [
              "null",
              "string"
            ]
          },
          "metadata": {
            "additionalProperties": {},
            "type": [
              "object",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "owner_id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "owner_id",
          "name",
          "slug",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
      "OrganizationInvitation": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "inviter_id": {
            "type": "string"
          },
          "organization_id": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/OrganizationInvitationStatus"
          }
        },
        "required": [
          "id",
          "email",
          "inviter_id",
          "organization_id",
          "role",
          "status",
          "expires_at",
          "created_at"
        ],
        "type": "object"
      },
      "OrganizationInvitationStatus": {
        "description": "The type of the invitation status",
        "enum": [
          "pending",
          "accepted",
          "rejected",
          "revoked",
          "expired"
        ],
        "type": "string"
      },
...
      "OrganizationSummary": {
        "properties": {
          "id": {
            "type": "string"
          },
          "logo": {
            "type": [
              "null",
              "string"
            ]
          },
          "metadata": {
            "additionalProperties": {},
            "type": [
              "object",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "owner_id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "owner_id",
          "name",
          "slug"
        ],
        "type": "object"
      },
...
      "Pagination": {
        "properties": {
          "has_more": {
            "type": "boolean"
          },
          "limit": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "total_pages": {
            "type": "integer"
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "total_pages",
          "has_more"
        ],
        "type": "object"
      },
...
      "UpdateOrganizationRequest": {
        "properties": {
          "logo": {
            "type": [
              "null",
              "string"
            ]
          },
          "metadata": {
            "additionalProperties": {},
            "type": [
              "object",
              "null"
            ]
          },
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "slug": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "type": "object"
      },
...
    }
  }
}