# plugins/organizations/handlers/organization_handlers.go
package handlers

import (
	"net/http"

	"github.com/Authula/authula/core/pagination"
	"github.com/Authula/authula/models"
	orgconstants "github.com/Authula/authula/plugins/organizations/constants"
	"github.com/Authula/authula/plugins/organizations/types"
	orgusecases "github.com/Authula/authula/plugins/organizations/usecases"
	"github.com/Authula/authula/util"
)

type CreateOrganizationHandler struct {
	UseCases *orgusecases.UseCases
}

func (h *CreateOrganizationHandler) Handle() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		ctx := r.Context()
		reqCtx, _ := models.GetRequestContext(ctx)
		actor := reqCtx.Actor

		var request types.CreateOrganizationRequest
		if err := util.ParseJSON(r, &request); err != nil {
			reqCtx.SetJSONResponse(http.StatusUnprocessableEntity, map[string]any{"message": err.Error()})
			reqCtx.Handled = true
			return
		}
		if err := request.Validate(); err != nil {
			reqCtx.SetJSONResponse(http.StatusUnprocessableEntity, map[string]any{"message": err.Error()})
			reqCtx.Handled = true
			return
		}

		organization, err := h.UseCases.CreateOrganization(ctx, actor, request)
		if err != nil {
			orgconstants.HandleError(err, reqCtx)
			return
		}

		reqCtx.Values[models.ContextAccessControlAssignRole.String()] = &models.AccessControlAssignRoleContext{
			UserID:         actor.ID,
			RoleName:       request.Role,
			AssignerUserID: nil,
		}

		reqCtx.SetJSONResponse(http.StatusCreated, organization)
	}
}

type GetAllOrganizationsHandler struct {
	UseCases *orgusecases.UseCases
}

func (h *GetAllOrganizationsHandler) Handle() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		ctx := r.Context()
		reqCtx, _ := models.GetRequestContext(ctx)
		actor := reqCtx.Actor
		organizations, err := h.UseCases.GetAllOrganizations(ctx, actor, pagination.ParseFromRequest(r))
		if err != nil {
			orgconstants.HandleError(err, reqCtx)
			return
		}

		reqCtx.SetJSONResponse(http.StatusOK, organizations)
	}
}

type GetOrganizationByIDHandler struct {
	UseCases *orgusecases.UseCases
}

func (h *GetOrganizationByIDHandler) Handle() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		ctx := r.Context()
		reqCtx, _ := models.GetRequestContext(ctx)
		actor := reqCtx.Actor

		organizationID := r.PathValue("organization_id")
		organization, err := h.UseCases.GetOrganizationByID(ctx, actor, organizationID)
		if err != nil {
			orgconstants.HandleError(err, reqCtx)
			return
		}

		reqCtx.SetJSONResponse(http.StatusOK, organization)
	}
}

type UpdateOrganizationHandler struct {
	UseCases *orgusecases.UseCases
}

func (h *UpdateOrganizationHandler) Handle() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		ctx := r.Context()
		reqCtx, _ := models.GetRequestContext(ctx)
		actor := reqCtx.Actor

		organizationID := r.PathValue("organization_id")

		var request types.UpdateOrganizationRequest
		if err := util.ParseJSON(r, &request); err != nil {
			reqCtx.SetJSONResponse(http.StatusUnprocessableEntity, map[string]any{"message": err.Error()})
			reqCtx.Handled = true
			return
		}
		if err := request.Validate(); err != nil {
			reqCtx.SetJSONResponse(http.StatusUnprocessableEntity, map[string]any{"message": err.Error()})
			reqCtx.Handled = true
			return
		}

		organization, err := h.UseCases.UpdateOrganization(ctx, actor, organizationID, request)
		if err != nil {
			orgconstants.HandleError(err, reqCtx)
			return
		}

		reqCtx.SetJSONResponse(http.StatusOK, organization)
	}
}

type DeleteOrganizationHandler struct {
	UseCases *orgusecases.UseCases
}

func (h *DeleteOrganizationHandler) Handle() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		ctx := r.Context()
		reqCtx, _ := models.GetRequestContext(ctx)
		actor := reqCtx.Actor

		organizationID := r.PathValue("organization_id")
		if err := h.UseCases.DeleteOrganization(ctx, actor, organizationID); err != nil {
			orgconstants.HandleError(err, reqCtx)
			return
		}

		reqCtx.SetJSONResponse(http.StatusOK, types.DeleteOrganizationResponse{Message: "organization deleted"})
	}
}